- use the appropriate types everywhere in SimulationRuntime/c/   f2c defines subroutine as typedef /* Subroutine */ int (*S_fp)(...);   so use that as the type of the functions sent in to the solvers. - enable -Wall in MinGW makefiles and get rid of most of the -Wall warnings.
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@15840 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/simulation/simulation_input_xml.cpp
@@ -346,7 +346,7 @@ void read_input_xml(int argc, char **argv,
   {
     WARNING(LOG_SIMULATION, "Error, input data file does not match model.");
     INDENT(LOG_SIMULATION);
-    WARNING2(LOG_SIMULATION, "nx in setup file: %ld from model code: %d", nxchk, modelData->nStates);
+    WARNING2(LOG_SIMULATION, "nx in setup file: %ld from model code: %d", nxchk, (int)modelData->nStates);
     WARNING2(LOG_SIMULATION, "ny in setup file: %ld from model code: %ld", nychk, modelData->nVariablesReal - 2*modelData->nStates);
     WARNING2(LOG_SIMULATION, "np in setup file: %ld from model code: %ld", npchk, modelData->nParametersReal);
     WARNING2(LOG_SIMULATION, "npint in setup file: %ld from model code: %ld", npintchk, modelData->nParametersInteger);
